+++ src/app/modules/layer-constellations/models/theme/constellations-style.ts
@@ -1,10 +1,11 @@
+import { LayerStyle } from '#core/models/theme/layer-style';
 import { LineStyle } from '#core/models/theme/line-style';
 import { TextStyle } from '#core/models/theme/text-style';
 
 /**
  * Defines the style properties of the constellations layer.
  */
-export interface ConstellationStyle {
+export interface ConstellationsStyle extends LayerStyle {
   /**
    * The properties of the lines representing the boundaries separating the constellations.
    */